requirements.txt
# Core dependencies
streamlit>=1.33.0
crewai>=0.54.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.30.0
python-dotenv>=1.0.0

# PDF generation
reportlab>=4.0.0

# HTTP requests
requests>=2.32.0

